+++ src/lib/Backend/Pocket/types/Pocket.d.ts
@@ -28,9 +28,7 @@ export type PocketArticle = {
 export type PocketGetResults = {
     status: number,
     complete: number,
-    list: {
-        [key: string]: PocketArticle
-    },
+    list: Record<string, PocketArticle>
     error: string | null,
     search_meta: { search_type: string },
     since: number
@@ -63,7 +61,7 @@ export type PocketAddResults = {
         used_fallback: string,
         lang: string,
         time_first_parsed: string,
-        authors: { [key: string]: [Object] },
+        authors: Record<string, Object[]>,
         top_image_url: string,
         resolved_normal_url: string,
         domain_metadata: {
@@ -76,6 +74,4 @@ export type PocketAddResults = {
     status: 1
 }
 
-export interface PocketRateLimit {
-    [key: string]: number
-}
+export type PocketRateLimit = Record<string, number>