+++ src/lib/Utility/SanitizeCommand.ts
@@ -15,7 +15,9 @@ const basic = new Permissions([
  * @param message 
  */
 export const Sanitize = (message: Message) => {
-    if(message.author.bot) {
+    if(message.webhookID) { // author is null in webhook messages
+        return false;
+    } else if(message.author.bot) {
         return false;
     } else if(message.type !== 'DEFAULT') {
         return false;